refactor: move to the event the routing mode, do not restrist to Async the effect
