+++ application/src/main/scala/io/github/positionpal/location/application/geo/MapsService.scala
@@ -4,11 +4,7 @@ import scala.concurrent.duration.FiniteDuration
 
 import cats.effect.Async
 import io.github.positionpal.location.commons.CanRaise
-import io.github.positionpal.location.domain.GPSLocation
-
-/** The mode of routing to a destination. */
-enum RoutingMode:
-  case Driving, Walking, Cycling
+import io.github.positionpal.location.domain.{GPSLocation, RoutingMode}
 
 /** An alias for the map service error. */
 type MapsServiceError = String
