+++ application/src/main/scala/io/github/positionpal/location/application/reactions/TrackingEventReactions.scala
@@ -1,7 +1,6 @@
 package io.github.positionpal.location.application.reactions
 
-import io.github.positionpal.location.application.geo.RoutingMode.Driving
-import io.github.positionpal.location.application.geo.{Distance, MapsService}
+import cats.Monad
 import io.github.positionpal.location.domain.{DrivingEvents, Route}
 
 /** A reaction to [[DrivingEvents.TrackingEvent]]s. */
@@ -16,17 +15,17 @@ object TrackingEventReaction extends BinaryShortCircuitReaction:
 
 /** A [[TrackingEventReaction]] checking if the position curried by the event is near the arrival position. */
 object IsArrivedCheck:
-  import TrackingEventReaction.*
-  import cats.effect.Async
   import cats.implicits.toFunctorOps
   import io.github.positionpal.location.application.geo.Distance.*
+  import io.github.positionpal.location.application.geo.MapsService
+  import TrackingEventReaction.*
 
   private val arrivalThreshold = 50.meters
 
-  def apply[M[_]: Async](mapsService: MapsService[M]): EventReaction[M] =
+  def apply[M[_]: Monad](mapsService: MapsService[M]): EventReaction[M] =
     on[M]: (route, event) =>
       for
-        distance: Distance <- mapsService.distance(Driving)(event.position, route.event.arrivalPosition)
+        distance <- mapsService.distance(route.sourceEvent.mode)(event.position, route.sourceEvent.arrivalPosition)
         outcome =
           if distance.toMeters.value <= arrivalThreshold.toMeters.value
           then Left(TriggerNotification("Arrived"))
@@ -36,11 +35,10 @@ object IsArrivedCheck:
 object IsContinuallyInSameLocationCheck:
   import TrackingEventReaction.*
   import cats.Monad
-  import cats.effect.Async
 
   private val samplingWindow = 5
 
-  def apply[M[_]: Async](): EventReaction[M] = on[M]: (route, event) =>
+  def apply[M[_]: Monad](): EventReaction[M] = on[M]: (route, event) =>
     if route.positions.take(samplingWindow).forall(_ == event.position)
     then Monad[M].pure(Left(TriggerNotification("Stuck")))
     else Monad[M].pure(Right(Continue))
