+++ application/src/test/scala/io/github/positionpal/location/application/reactions/TrackingEventReactionsTest.scala
@@ -1,5 +1,6 @@
 package io.github.positionpal.location.application.reactions
 
+import io.github.positionpal.location.domain.RoutingMode
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 
@@ -13,7 +14,9 @@ class TrackingEventReactionsTest extends AnyFunSpec with Matchers:
 
   import java.util.Date
 
-  private val route = Route(DrivingEvents.StartRoutingEvent(Date(), UserId("test"), GPSLocation(0.0, 0.0)))
+  private val route = Route(
+    DrivingEvents.StartRoutingEvent(Date(), UserId("test"), RoutingMode.Driving, GPSLocation(0.0, 0.0)),
+  )
   private val event: TrackingEvent = TrackingEvent(Date(), UserId("test"), GPSLocation(0.1, 0.1))
 
   describe("`TrackingEventReaction`s"):
