+++ infrastructure/src/main/scala/io/github/positionpal/location/infrastructure/geo/MapsServiceAdapter.scala
@@ -7,9 +7,9 @@ import cats.effect.IO
 import cats.implicits.catsSyntaxEither
 import io.circe.Json
 import io.github.positionpal.location.application.geo.Distance.meters
-import io.github.positionpal.location.application.geo.RoutingMode.Driving
-import io.github.positionpal.location.application.geo.{Distance, MapsService, MapsServiceError, RoutingMode}
-import io.github.positionpal.location.domain.GPSLocation
+import io.github.positionpal.location.application.geo.{Distance, MapsService, MapsServiceError}
+import io.github.positionpal.location.domain.RoutingMode.Driving
+import io.github.positionpal.location.domain.{GPSLocation, RoutingMode}
 import org.http4s.Uri
 import org.http4s.circe.jsonOf
 import org.http4s.client.Client
